Add check for existing files in sysroot
Now the error for overwriting files in sysroot is not printed when trying
to copy files which are not already there. Before if there was any file in
sysroot, the build have failed. The files printed with error are also
just the files which were actually tried to copy into sysroot, not all
files in sysroot.